backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
